+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ Change Log
 - [CHANGED] update create_collection docstrings
 - [CHANGED] update HV/MV transformer standard type data
 - [ADDED] pp_elements() toolbox function
+- [ADDED] new parameter g_us_per_km to model dielectric losses in lines
+- [ADDED] single phase short-circuit calculation with negative sequence models
+- [ADDED] generic storage model (sgen/load like element with negative / positive power allowed)
 
 [1.4.3] - 2018-02-06
 ----------------------